+++ src/internal/MicrogrammarSpecParser.ts
@@ -1,6 +1,6 @@
-import {Concat, toMatchingLogic} from "../Concat";
 import {Config, DefaultConfig} from "../Config";
 import {MatchingLogic} from "../Matchers";
+import {Concat, toMatchingLogic} from "../matchers/Concat";
 import {isPatternMatch} from "../PatternMatch";
 import {Literal} from "../Primitives";
 import * as MatcherPrinter from "./MatcherPrinter";